// perforator/ui/src/generated/perforator/proto/perforator/perforator.ts
export interface SymbolizeOptions {
  /**
   * Convert raw instruction addresses to the function names & other metadata.
   * Super slow.
   * Enabled by default.
   */
  Symbolize?:
    | boolean
    | undefined;
  /**
   * Convert raw function names to human-readable form.
   * Enabled by default.
   */
  Demangle?:
    | boolean
    | undefined;
  /**
   * Decode inlined functions.
   * Enabled by default.
   */
  Inlines?:
    | boolean
    | undefined;
  /**
   * Embed dwarf discriminator numbers into high 32 bits of the line numbers.
   * We do not control pprof.Profile format, so we cannot add discriminator field.
   */
  EmbedDwarfDiscriminators?: boolean | undefined;
}

export interface FlamegraphOptions {
  /** Truncate stacks deeper than @MaxDepth. */
  MaxDepth?:
    | number
    | undefined;
  /**
   * Minimum relative function weight to draw.
   * 1.0 means "full flamegraph".
   */
  MinWeight?:
    | number
    | undefined;
  /** Build flamegrph from the top. */
  Inverse?:
    | boolean
    | undefined;
  /** Show line numvers, if available */
  ShowLineNumbers?:
    | boolean
    | undefined;
  /**
   * Show file names after function names.
   * Enabled by default.
   */
  ShowFileNames?: boolean | undefined;
}

/** Empty */
export interface RawProfileOptions {
}

/** Empty */
export interface LLVMPGOProfileOptions {
}

export interface PostprocessOptions {
  MergePythonAndNativeStacks?: boolean | undefined;
}

export interface RenderFormat {
  /** How to symbolize the profile. */
  Symbolize: SymbolizeOptions | undefined;
  Postprocessing:
    | PostprocessOptions
    | undefined;
  /** Build flamegraph. */
  Flamegraph?:
    | FlamegraphOptions
    | undefined;
  /** Return raw profile, potentiatlly symbolized. */
  RawProfile?:
    | RawProfileOptions
    | undefined;
  /**
   * Return profile suitable for the LLVM PGO.
   * See https://clang.llvm.org/docs/UsersManual.html#sample-profile-text-format.
   * @deprecated
   *
   * @deprecated
   */
  LLVMPGOProfile?:
    | LLVMPGOProfileOptions
    | undefined;
  /**
   * Build flamegraph.
   * Json is parsed faster in the ui.
   * html can be shown without ui.
   */
  JSONFlamegraph?: FlamegraphOptions | undefined;
}